Roadmap Item Done ("SpeakerBuffer - Unignore SpeakerBufferTest.sequencing").
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@2953 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ sneer/bricks/src/sneer/skin/sound/mic/tests/MicTest.java
@@ -44,7 +44,7 @@ public class MicTest {
 	@Test
 	public void testPacketSequence() throws Exception {
 		
-		final HashMap<Short, PcmSoundPacket> seenPackets = new HashMap<Short, PcmSoundPacket>();
+		final HashMap<Integer, PcmSoundPacket> seenPackets = new HashMap<Integer, PcmSoundPacket>();
 		tupleSpace.addSubscription(PcmSoundPacket.class, new Omnivore<PcmSoundPacket>() { @Override public void consume(PcmSoundPacket packet) {
 			seenPackets.put(packet.sequence, packet);
 		}});
